requirements.txt
# Core dependencies
llama-index>=0.10.0
llama-index-llms-huggingface>=0.2.0
llama-index-embeddings-huggingface>=0.2.0
transformers>=4.44.0
torch>=2.0.0
accelerate>=0.20.0
bitsandbytes>=0.41.0

# Web interface
streamlit>=1.28.0
streamlit-chat>=0.1.1
plotly>=5.17.0

# PDF generation
reportlab>=4.0.0
fpdf2>=2.7.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Vector database
chromadb>=0.4.0
sentence-transformers>=2.2.0
llama-index-vector-stores-chroma>=0.1.0

# PDF processing for tax documents
pypdf>=3.15.0
pdfplumber>=0.9.0

# Document processing for Indian tax documents
openpyxl>=3.1.0
xlrd>=2.0.0
tabula-py>=2.8.0
camelot-py[cv]>=0.11.0

# OCR for investment receipts and bills
pytesseract>=0.3.10
Pillow>=10.0.0

# Additional utilities
requests>=2.31.0
python-dateutil>=2.8.0
pathlib>=1.0.0

# Google Drive integration
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
google-auth>=2.23.0

# Document classification and AI
scikit-learn>=1.3.0
opencv-python>=4.8.0

# Indian tax specific libraries

# Environment and logging
python-dotenv>=1.0.0
loguru>=0.7.0